backend/requirements.txt
aiohttp==3.7.4
aiohttp-jwt==0.6.1
aiohttp-middlewares==1.1.0
psycopg2==2.9.1
PyJWT==2.1.0
SQLAlchemy==1.4.11
